Add basic access control

+++ browse.php
@@ -9,6 +9,8 @@ $iterator = new RecursiveIteratorIterator($dir_iterator, RecursiveIteratorIterat
 // could use CHILD_FIRST if you so wish
 //
 
+$ownedScripts = getSessionScripts();
+
 $count = 0;
 echo "<table>";
 foreach ($iterator as $file) {
@@ -20,8 +22,11 @@ foreach ($iterator as $file) {
     $name = $file->getPath()."/".$file->getFilename();
     $name = preg_replace("/^".SAVE_DIR."\//", "", $name);
     echo "<tr></td>";
-    echo "<td class='script'><a class='xlscript' href='index.php?name=$name'>$name</a></td>";
-    echo "<td>[<a class='del' href='delete.php?name=$name'>delete</a>]</td>";
+    echo "<td class='script'><a class='xlscript' href='index.php?name=$name'>";
+    echo htmlspecialchars($name);
+    echo "</a></td>";
+    if (@$ownedScripts[$name])
+        echo "<td>[<a class='del' href='delete.php?name=$name'>delete</a>]</td>";
     echo "</tr>";
 }
 echo "</table>";
